+++ CHANGELOG.md
@@ -234,6 +234,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - New `MovableMan` Lua function, `KillAllTeamActors`, which kills all `Actor`s on the given `Team`.
 
+- New `ACRocket` INI property `MaxGimbalAngle`, which enables automatic stabilization via tilting of the main engine.
+
 </details>
 
 <details><summary><b>Changed</b></summary>